# back-end/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Bitcoin HD and n-out-of-m Multisignature P2SH Address Generator",
    "version": "1.0"
  },
  "paths": {
    "/segwit_address_generator/{seedPhrase}/{path}": {
      "get": {
        "tags": [
          "HD Segwit Bitcoin Address"
        ],
        "description": "Hierarchical Deterministic (HD) Segregated Witness (SegWit) bitcoin address from given seed phrase and path",
        "parameters": [
          {
            "in": "path",
            "schema": {
              "type": "string"
            },
            "name": "seedPhrase",
            "required": true,
            "description": "Expecting a seed pharse to derive seed",
            "example": "glow laugh acquire menu anchor evil occur put hover renew calm purpose marry run speed"
          },
          {
            "in": "path",
            "schema": {
              "type": "string"
            },
            "name": "path",
            "required": true,
            "description": "Enter the derivation path",
            "example": "m/84'/0'/0'/0"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "bc1qrd8rth4r228a4qhthk9lf0yve9c90uh4uret3x"
                }
              }
            }
          },
          "400": {
            "description": "Resource Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Please check whether the request format is correct for both seed phrase and path"
                }
              }
            }
          },
          "412": {
            "description": "Incorrect Input Parameters",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Please check the input parameters"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Please try again if the server is active"
                }
              }
            }
          }
        }
      }
    },
    "/multisig/{m}/{n}/{Public Keys}": {
      "get": {
        "tags": [
          "n-out-of-m Multisignature (multi-sig) Pay-To-Script-Hash (P2SH) bitcoin address"
        ],
        "description": "Requires n,m and public keys, Important to note that n<m",
        "parameters": [
          {
            "in": "path",
            "schema": {
              "type": "integer"
            },
            "name": "m",
            "required": true,
            "description": "m value for n-out-of-m",
            "example": 2
          },
          {
            "in": "path",
            "schema": {
              "type": "integer"
            },
            "name": "n",
            "required": true,
            "description": "n value for n-out-of-m",
            "example": 3
          },
          {
            "in": "path",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "name": "Public Keys",
            "required": true,
            "description": "Array of length n for public keys",
            "example": [
              "02e135e12b4417b41e92e738448cb51581c70f14bf885b0d4056ac2c3cc5c8729c",
              "022015b568fb0f2f792e2e1d230a7f64e8a75b5d4a3ae549b55c3724cdc148b32c",
              "02799dc04a8acf04e793ff0f2c35c20c0388529eb964c565a455f13c07123c9ea2"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Resource Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Please check whether the request format is correct for both seed phrase and path"
                }
              }
            }
          },
          "412": {
            "description": "Incorrect Input Parameters",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Please check the input parameters"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Please try again if the server is active"
                }
              }
            }
          }
        }
      }
    }
  }
}